requirements.txt
streamlit
transformers
torch
Pillow
accelerate
tf-keras